+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # AUX Changelog
 
+## V0.5.4
+
+### Date: 04/29/2019
+
+### Changes:
+
+-   Improvements
+    -   Changed AUX Player's default background color to match the dark background color that AUX Builder uses.
+    -   Changed the globals file to look like a normal file when created and be labeled as "Global".
+    -   Updated all the formula functions to use the new naming scheme.
+    -   Added the ability to drag worksurfaces when they are minimized.
+        -           Setting `aux.context.movable` to `false` will prevent this behavior.
+    -   Selecting an item in the inventory no longer shows a selection indicator.
+-   Bug Fixes
+    -   The inventory placeholders should now always appear square.
+    -   Dragging an item out of the inventory will now always remove the image of that item in the inventory.
+
 ## V0.5.3
 
 ### Date: 04/26/2019